pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "disfun"
version = "0.1.0"
description = "DistributedFunSearch - Multi-node distributed implementation of FunSearch with cluster support via RabbitMQ"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "scipy",
  "transformers",
  "openai",
  "aio-pika",
  "yarl",
  "psutil",
  "GPUtil",
  "memory-profiler",
  "cloudpickle",
  "nvidia-ml-py",
  "accelerate",
  "lmdb",
  "python-Levenshtein",
  "tqdm",
  "wandb",
  "matplotlib",
  "pandas",
  "seaborn",
  "tabulate",
  "networkx"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
disfun = "disfun.__main__:main"
